+++ SharePointFramework/PortfolioWebParts/src/components/IdeaModule/Commands/Commands.module.scss
@@ -1,0 +1,7 @@
+.commands {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  height: 42px;
+  margin: 0px auto;
+}
