+++ SharePointFramework/PortfolioWebParts/src/components/IdeaModule/Commands/index.ts
@@ -1,0 +1,1 @@
+export * from './Commands'
