+++ SharePointFramework/PortfolioWebParts/src/components/IdeaModule/Commands/useDelete.ts
@@ -1,0 +1,14 @@
+import { useIdeaModuleContext } from '../context'
+
+/**
+ * Hook for deletion of idea. Returns a callback function
+ * for deleting the selected report.
+ *
+ * @returns A function callback
+ */
+export const useDelete = () => {
+  const context = useIdeaModuleContext()
+  return async (): Promise<void> => {
+    console.log('delete', context)
+  }
+}
