+++ SharePointFramework/PortfolioWebParts/src/components/IdeaModule/IdeaModule.module.scss
@@ -18,7 +18,7 @@
       flex-direction: column;
       gap: 16px;
 
-      .toolbar {
+      .phasebar {
         display: flex;
         flex-direction: column;
         gap: 16px;
@@ -42,7 +42,6 @@
     .accordion {
       width: fit-content;
       padding: 0 8px;
-      margin-bottom: 16px;
 
       &:hover {
         background-color: var(--colorNeutralBackground1Hover);
